Customer-Portal/requirements.txt
bcrypt==4.0.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
requests>=2.31.0
werkzeug>=2.3.8
pytest-env==1.0.1
pyOtp==2.9.0
Flask-JWT-Extended==4.5.3
qrcode==7.4.2
pyOpenSSL==23.2.0